refactor(requests): extract RequestLifecycleService and shared status constants
Pull every request-shaped method out of MediaService into a new
RequestLifecycleService that owns the cross-module orchestration. The
state machine for a request — `APPROVED → PROCESSING → AVAILABLE` plus
the admin-driven `DECLINED` shortcuts — now lives in one file and
reads top-to-bottom.

What moved out of MediaService

Seven methods (~250 lines) deleted from media.service.ts:
  linkOpenRequestsToImportedMedia
  declineLinkedActiveRequests
  declineLinkedActiveRequestsForSeason
  onReleaseGrabbedForMedia
  onMediaFilesChanged
  isRequestScopeCovered
  OnModuleInit/Destroy + EventsService subscription

Plus the EventsService and NotificationsService injections that only
existed to support those methods, and the FliksRequest entity from
TypeOrmModule.forFeature.

What stays on MediaService

Just the narrow public surface RequestLifecycleService calls back into:
  - applyMonitoredForRequestScope (mutates media.monitored /
    season.monitored — genuinely a media op).
  - findOneWithSeasonsAndFiles (centralised relation-eager load used
    by the coverage check).
  - findByTmdbId (already existed).
The update / updateSeason / remove paths now delegate to
requestLifecycle.onMediaMonitorChange / onSeasonMonitorChange /
onMediaRemoved with the same effect.

Shared building blocks

request-status.constants.ts collects:
  - ACTIVE_REQUEST_STATUSES (PENDING/APPROVED/PROCESSING/AVAILABLE)
  - SATISFIABLE_REQUEST_STATUSES (APPROVED/PROCESSING/AVAILABLE)
  - IN_FLIGHT_REQUEST_STATUSES (PENDING/APPROVED/PROCESSING)
  - seasonScopeOf(request) — normalises `null/[]/[1,2]` into
    `null | Set<number>`.

RequestsService now imports the two former constants from there and
the new helper. The `coversSeasons` check is one branch shorter.

Wiring

  RequestsModule provides + exports RequestLifecycleService and now
  imports MediaModule via forwardRef.
  MediaModule imports RequestsModule via forwardRef so MediaService
  can inject RequestLifecycleService.
  AutoGrabPipelineService swaps its forwardRef(MediaService) for
  forwardRef(RequestLifecycleService) — the hook gets called on the
  service that actually owns it.

No behaviour change: every external entry point (controllers, scheduler
RSS / missing-episode scans, completion.service via import.complete)
still triggers the same downstream effects. Type-check clean.
