+++ backend/src/modules/requests/requests.module.ts
@@ -1,4 +1,4 @@
-import { Module } from '@nestjs/common';
+import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { FliksRequest } from './entities/request.entity';
 import { RequestComment } from './entities/request-comment.entity';
@@ -8,6 +8,7 @@ import { NotificationsModule } from '../notifications/notifications.module';
 import { MediaModule } from '../media/media.module';
 import { ProfilesModule } from '../profiles/profiles.module';
 import { RequestsService } from './requests.service';
+import { RequestLifecycleService } from './request-lifecycle.service';
 import { RequestsController } from './requests.controller';
 import { AutoApprovalRulesController } from './auto-approval-rules.controller';
 
@@ -16,11 +17,14 @@ import { AutoApprovalRulesController } from './auto-approval-rules.controller';
     TypeOrmModule.forFeature([FliksRequest, RequestComment, AutoApprovalRule]),
     AuthModule,
     NotificationsModule,
-    MediaModule,
+    // forwardRef: MediaModule injects RequestLifecycleService for the
+    // import / remove / unmonitor hooks, and we inject MediaService here
+    // for the monitoring + 409 fallback lookups.
+    forwardRef(() => MediaModule),
     ProfilesModule,
   ],
   controllers: [RequestsController, AutoApprovalRulesController],
-  providers: [RequestsService],
-  exports: [RequestsService],
+  providers: [RequestsService, RequestLifecycleService],
+  exports: [RequestsService, RequestLifecycleService],
 })
 export class RequestsModule {}
